+++ foxy_farmer/monkey_patch_chia_version.py
@@ -1,8 +1,8 @@
 import chia
-import pkg_resources
+
+from foxy_farmer.version import version
 
 chia_version = chia.__version__
-version = pkg_resources.require("foxy-farmer")[0].version
 
 
 def monkey_patch_chia_version():